Remove suspense boundary from token refresh.
